refactor frisbeegen file structure
