+++ go.mod
@@ -9,6 +9,6 @@ require (
 	github.com/nats-io/nats-server/v2 v2.2.0 // indirect
 	github.com/nats-io/nats.go v1.10.1-0.20210228004050-ed743748acac
 	github.com/pkg/errors v0.9.1
-	github.com/rs/zerolog v1.20.0
+	github.com/rs/zerolog v1.22.0
 	github.com/stretchr/testify v1.7.0
 )
