+++ internal/utils/stringUtils.go
