--- protoc-gen-go-frisbeegen/go.mod
@@ -1,8 +1,0 @@
-module protoc-gen-go-frisbeegen
-
-go 1.16
-
-require (
-	github.com/golang/protobuf v1.5.2
-	google.golang.org/protobuf v1.26.0
-)
